requirements.txt
discord.py==2.4.0
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
Flask
